requirements.txt
fastapi
uvicorn
pydantic
roman
strawberry-graphql